fix: compile errors - remove parcelize import, fix shl syntax, Long->Int for limitToLast, animateInt->animateFloat

+++ app/src/main/java/com/dedroneTECTION/repository/FirebaseRepository.kt
@@ -22,7 +22,7 @@ class FirebaseRepository {
         }
     }
 
-    suspend fun getRecentDetections(limit: Long = 50): List<Map<String, Any?>> {
+    suspend fun getRecentDetections(limit: Int = 50): List<Map<String, Any?>> {
         return try {
             val snapshot = detectionsRef
                 .orderByChild("timestamp")
